Move _shuffle_arrays() to _utils.py
